requirements.txt
requests==2.32.3
Flask-APScheduler==1.13.1
pillow==10.4.0
beautifulsoup4==4.12.3
pydub==0.25.1
Pyrebase4==4.8.0